Renamed CaloTowerFwd.h to CaloTowerDefs.h
Because CaloTowerFwd.h is not really forward declaring anything, we
rename it to CaloTowerDefs.h to better represent it's purpose.
