+++ DataFormats/CaloTowers/interface/CaloTowerCollection.h
@@ -2,6 +2,6 @@
 #define DATAFORMATS_CALOTOWERS_CALOTOWERCOLLECTION_H 1
 
 #include "DataFormats/CaloTowers/interface/CaloTower.h"
-#include "DataFormats/CaloTowers/interface/CaloTowerFwd.h"
+#include "DataFormats/CaloTowers/interface/CaloTowerDefs.h"
 
 #endif
